RTM: update CreateScenarios utility to take a single base scenario
Now that all data about a scenario is in one, modify the model to create
the needed AM, MD and PM networks from a single source scenario.

This routine copies the base scenario and sets the master attirbutes for
lanes, vdf, tpf and hdwy to the specific time of day value. mode=v is then
removed from all lanes=0 links and all transit lines with a zero headway
are deleted.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>