RTM: consolidate a block of the truck model aggregation code
Consolidate a single speclist for all PCE and market aggregation
calculations. Remove an intermediate compute matrix that can be done in a
single calculation. Remove a loop over matrix lists and directly add the
matrix specs as documentation can now be added for which trucks/time of
day applies to each block.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>