RTM: update auto assignment to skim each op cost separately
Skim the fixed costs portion of the generalized cost per class. Use the
perclass fixed cost to calculate actual travel time minutes and distance
by subtracting the weighted fixed costs form the generalized cost minutes
matrix. Calculate the distance matrix by subtracting the existing toll
skim from the fixed cost and dividing by the vehicle op cost.

This distance is slightly incorrect for the LGV and HGV classes due to the
extra toll weighting and @tkpen values, but is not important for the
current formulation of the model.

Calculate the intrazonal estimates from the individual skims rather than
the GC skims.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>