RTM: correct screenline tagging for intramunicipal screenlines
Screenline 4,5,11,16 had batchin issues with the updated network and were
not fixed by the previous renumbering.

Note that screenlines 5 and 11 are missing a tag for the WCE as this link
is tagged in screenline 4 and cannot belong to multiple screenlines.

Some highway 1 link tagging was also fixed in the future years containing
the gateway upgrades for screenlines 5, 11 and 16.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>