RTM: add new ttf type 20 for bus links at a signal
Bus only lanes show very little congestion and are essentially free-flow
links, add a new ttf to represent a link with a signalized control that
adds a fixed 20 second overhead to the travel time.

This value is based on the accel/decel delay assumed in the new
boarding/alighting based delay added in Phase 3.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>