RTM: change signalized intersection vdfs to include a fixed delay component
Use ul1 as a fixed delay of 15 seconds (0.25 minutes) in the MD time
period only for now. This will slow the average network speeds down to
something much more reasonable.

It is likely that further refinement of the vdf formulations would allow
this to be used in the AM and PM time periods as well, but might require
adjustment to the alpha and beta coefficients to make them less steep to
offset the new fixed delay.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>