RTM: add initial network calculation overide functionality
If a file called custom_network.txt is found in the Inputs/ folder during
a model run, it will be parsed to perform additional link calculations on
each of the time of day networks being created from the master scenario.

These can be arbitrary calculations as the arguments are split
comma-delimited and fed to the util.emme_link_calc() helper method for
evaluation on the appropriate scenario.

leading # characters will comment the entire line allowing for comments as
to the purpose of the calcualtion.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>