RTM: add optional toll skim functionality to run_model
Update the assignment procedure to optionally run a toll skim explicitly.

The format of the path analysis has been made common to all a
append()-based version easier to implement to run the toll skim and opcost
skim together in the same assignment, currently this produces an error in
the SOLA toll that is being investigated.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>