RTM: limit number of averaging iterations in transit assignment
When congestion is disabled to generate the initial skims, limit the
iterations to 2 as we're assigning a zero matirx anyway and this adds
significant runtime to the early model stages.

Avoid passing state by writing the self.max_iterations value as reusing
the transit assignment object can have unexpected results if this state
is overwritten, use a local variable in the assignment method.

Signed-off-by: Harvey Harrison <hharrison@mcelhanney.com>