RTM: add a per-segment effective boarding penalty
Currently all segments are calculated to equal the default value defined
at the transit line level. This can be further modified by a custom
network file setting segment values.

@brdline - Line Boarding Penalty in minutes (default 1)
@brdpeneff - The effective boarding penalty on transit segments

@brdpeneff is the value actually skimmed by the transit assignment into
the boardings matrices, this required a change to the bus skimming
which was collecting the number of boardings instead of the boarding
penalty. This used to be equivalent with the 1 minute boarding value,
but has been changed to the approach used by Rail and WCE now.